requirementsts.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
Django==5.1.5
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-tables2==2.7.5
djangorestframework==3.15.2
idna==3.10
mysqlclient==2.2.7
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
